refactor: layout 수정
