+++ src/router.jsx
@@ -1,17 +1,16 @@
 import { createBrowserRouter } from 'react-router-dom';
-import App from './App';
 import Main from './pages/main/Main';
 import NotFoundError from './pages/error/NotFound';
 import Login from './pages/login/Login';
 import Signup from './pages/signup/Signup';
 import Chart from './pages/chart/Chart';
 import Chatting from './pages/chatting/Chatting';
-import RootLayout from './layout/rootLayout';
+import App from './App';
 
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <RootLayout />,
+    element: <App />,
     children: [
       {
         path: '',
